+++ src/data/GraphProblems.kt
@@ -67,6 +67,28 @@ data class GraphPartition(
 	}
 }
 
+data class GraphPath(
+	val vertices: List<Vertex> = emptyList(),
+	val distance: List<List<Double>> = emptyList(),
+) {
+	override fun toString(): String = "Vertices: $vertices\nDistance: $distance"
+
+	companion object {
+		fun new(name: String): GraphPath? {
+			val data = load(name) ?: return null
+			val vertexList = data["vertices"].toStringList()
+			val distanceMatrix = mutableListOf<List<Double>>()
+			for(line in data["distance"].parseList()) {
+				distanceMatrix.add(line.matrixRow(true))
+			}
+			return GraphPath(
+				vertices = vertexList,
+				distance = distanceMatrix,
+			)
+		}
+	}
+}
+
 typealias GraphVariablesFn = (Graph) -> List<String>
 typealias GraphSpanFn = (GraphCfg) -> List<String>
 typealias GraphColorsFn<T> = (GraphColoring) -> List<T>